refactor(generator/rust): split `Generate()` function
This organizes the code a little better, moving the main public API to
its own file (module a really small number of helpers).
